% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faRating.R
\name{faRatingInput}
\alias{faRatingInput}
\title{Shiny Rating Input}
\usage{
faRatingInput(inputId, default = NULL, configuration = list())
}
\arguments{
\item{inputId}{\code{character} id for the input}

\item{default}{\code{integer} value for the initial rating value}

\item{configuration}{\code{list} of initial 'props' for the rating input}
}
\value{
\code{shiny input}
}
\description{
Input to provide an interface for a user to rate something often indicated by a star
}
