% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faSlider.R
\name{faSliderInput}
\alias{faSliderInput}
\title{Shiny Slider Input}
\usage{
faSliderInput(inputId, default = NULL, configuration = NULL)
}
\arguments{
\item{inputId}{\code{character} id for the input}

\item{default}{\code{integer} value for the initial Slider value}

\item{configuration}{\code{list} of initial 'props' for the Slider input}
}
\value{
\code{shiny input}
}
\description{
Input to provide an interface for a user to rate something often indicated by a star
}
