% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faSpinbutton.R
\name{faSpinbuttonInput}
\alias{faSpinbuttonInput}
\title{Shiny Spinbutton Input}
\usage{
faSpinbuttonInput(inputId, default = NULL, configuration = NULL)
}
\arguments{
\item{inputId}{\code{character} id for the input}

\item{default}{\code{integer} value for the initial spin button value}

\item{configuration}{\code{list} of initial 'props' for the spin button input}
}
\value{
\code{shiny input}
}
\description{
Input to provide an interface for a user to rate something often indicated by a star
}
