% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faToggle.R
\name{faToggleInput}
\alias{faToggleInput}
\title{Shiny Toggle Input}
\usage{
faToggleInput(inputId, default = NULL, configuration = NULL)
}
\arguments{
\item{inputId}{\code{character} id for the input}

\item{default}{\code{integer} value for the initial toggle value}

\item{configuration}{\code{list} of initial 'props' for the toggle input}
}
\value{
\code{shiny input}
}
\description{
Input to provide an interface for a user to rate something often indicated by a star
}
