% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faCheckbox.R
\name{updateFaCheckboxInput}
\alias{updateFaCheckboxInput}
\title{Update Shiny Checkbox Input}
\usage{
updateFaCheckboxInput(session, inputId, value = NULL,
  configuration = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{inputId}{\code{character} name/id for the shiny input checkbox to change}

\item{value}{\code{integer} value to assign}

\item{configuration}{\code{list} of 'props' to change the checkbox}
}
\description{
Allow changing or updating the configuration or value of a shiny checkbox input.
}
