% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faSlider.R
\name{updateFaSliderInput}
\alias{updateFaSliderInput}
\title{Update Shiny Slider Input}
\usage{
updateFaSliderInput(session, inputId, value = NULL,
  configuration = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{inputId}{\code{character} name/id for the shiny input Slider to change}

\item{value}{\code{integer} value to assign}

\item{configuration}{\code{list} of 'props' to change the Slider}
}
\description{
Allow changing or updating the configuration or value of a shiny Slider input.
}
