% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faSpinbutton.R
\name{updateFaSpinbuttonInput}
\alias{updateFaSpinbuttonInput}
\title{Update Shiny Spinbutton Input}
\usage{
updateFaSpinbuttonInput(session, inputId, value = NULL,
  configuration = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{inputId}{\code{character} name/id for the shiny input spin button to change}

\item{value}{\code{integer} value to assign}

\item{configuration}{\code{list} of 'props' to change the spin button}
}
\description{
Allow changing or updating the configuration or value of a shiny spin button input.
}
