% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faToggle.R
\name{updateFaToggleInput}
\alias{updateFaToggleInput}
\title{Update Shiny Toggle Input}
\usage{
updateFaToggleInput(session, inputId, value = NULL,
  configuration = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{inputId}{\code{character} name/id for the shiny input toggle to change}

\item{value}{\code{integer} value to assign}

\item{configuration}{\code{list} of 'props' to change the toggle}
}
\description{
Allow changing or updating the configuration or value of a shiny toggle input.
}
