% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faCheckbox.R
\name{faCheckboxInput}
\alias{faCheckboxInput}
\title{Shiny Checkbox Input}
\usage{
faCheckboxInput(inputId, default = FALSE)
}
\arguments{
\item{inputId}{\code{character} id for the input}

\item{default}{\code{logical} value for the initial checkbox value}
}
\value{
\code{shiny input}
}
\description{
Input to provide an interface for a user to rate something often indicated by a star
}
