% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faRating.R
\name{faRatingInput}
\alias{faRatingInput}
\title{Shiny Rating Input}
\usage{
faRatingInput(inputId, default = "")
}
\value{
\code{shiny input}
}
\description{
Input to provide an interface for a user to rate something often indicated by a star
}
